requirements.txt
Django==3.1.7
djangorestframework==3.12.2
django-cors-headers==3.7.0
requests